requirements.txt
python-dateutil>=2.8.2
typing-extensions>=4.5.0
openai==1.12.0
supabase==1.2.0
streamlit==1.32.0
python-dotenv==1.0.1
pandas>=2.0.0
numpy>=1.24.0
cryptography>=42.0.0
requests==2.31.0 